modified enrolling and unenrolling students
= from the class list the delete link no longer deletes the user
from the system. Instead it unenrols the student from the course.

= an option is given for importing users. Instructors and admins
can now choose to merge their new class list with the class list
they have previously imported. Another option is to update the
class list and remove all students that were on the list in the
previous import but not on the new list.

+++ app/controllers/users_controller.php
@@ -817,10 +817,18 @@ class UsersController extends AppController
             $this->cakeError('error404');
         }
 
-        if ($this->User->delete($id)) {
-            $this->Session->setFlash(__('Record is successfully deleted!', true), 'good');
+        if (is_null($courseId)) {
+            if ($this->User->delete($id)) {
+                $this->Session->setFlash(__('Record is successfully deleted!', true), 'good');
+            } else {
+                $this->Session->setFlash(__('Error: Delete failed!', true));
+            }
         } else {
-            $this->Session->setFlash(__('Error: Delete failed!', true));
+            if ($this->User->removeStudent($id, $courseId)) {
+                $this->Session->setFlash(__('Student is successfully unenrolled!', true), 'good');
+            } else {
+                $this->Session->setFlash(__('Error: Unenrol failed!', true));
+            }
         }
 
         $this->redirect($this->referer());
@@ -977,6 +985,7 @@ class UsersController extends AppController
             }
 
             $data = Toolkit::parseCSV($uploadFile);
+            $usernames = array();
             // generation password for users who weren't given one
             foreach ($data as &$user) {
                 if (empty($user[User::IMPORT_PASSWORD])) {
@@ -984,6 +993,11 @@ class UsersController extends AppController
                 } else {
                     $user[User::GENERATED_PASSWORD] = '';
                 }
+                $usernames[] = $user[User::IMPORT_USERNAME];
+            }
+            
+            if ($this->data['User']['update_class']) {
+                $this->User->removeOldStudents($usernames, $courseId);
             }
 
             // add the users to the database
